+++ src/OpenTelemetry.Instrumentation.Wcf/Implementation/ClientChannelInstrumentation.cs
@@ -51,7 +51,7 @@ internal static class ClientChannelInstrumentation
 
                 if (WcfInstrumentationActivitySource.Options!.SetSoapMessageVersion)
                 {
-                    activity.SetTag(WcfInstrumentationConstants.SoapMessageVersionTag, request.Version.ToString());
+                    activity.SetTag(WcfInstrumentationConstants.AttributeSoapMessageVersion, request.Version.ToString());
                 }
 
                 var remoteAddressUri = request.Headers.To ?? remoteChannelAddress;
@@ -59,13 +59,13 @@ internal static class ClientChannelInstrumentation
                 {
                     activity.SetTag(SemanticConventions.AttributeServerAddress, remoteAddressUri.Host);
                     activity.SetTag(SemanticConventions.AttributeServerPort, remoteAddressUri.Port);
-                    activity.SetTag(WcfInstrumentationConstants.WcfChannelSchemeTag, remoteAddressUri.Scheme);
-                    activity.SetTag(WcfInstrumentationConstants.WcfChannelPathTag, remoteAddressUri.LocalPath);
+                    activity.SetTag(WcfInstrumentationConstants.AttributeWcfChannelScheme, remoteAddressUri.Scheme);
+                    activity.SetTag(WcfInstrumentationConstants.AttributeWcfChannelPath, remoteAddressUri.LocalPath);
                 }
 
                 if (request.Properties.Via != null)
                 {
-                    activity.SetTag(WcfInstrumentationConstants.SoapViaTag, request.Properties.Via.ToString());
+                    activity.SetTag(WcfInstrumentationConstants.AttributeSoapVia, request.Properties.Via.ToString());
                 }
 
                 try
@@ -106,7 +106,7 @@ internal static class ClientChannelInstrumentation
 
                 if (reply != null)
                 {
-                    activity.SetTag(WcfInstrumentationConstants.SoapReplyActionTag, reply.Headers.Action);
+                    activity.SetTag(WcfInstrumentationConstants.AttributeSoapReplyAction, reply.Headers.Action);
                     try
                     {
                         WcfInstrumentationActivitySource.Options!.Enrich?.Invoke(activity, WcfEnrichEventNames.AfterReceiveReply, reply);
@@ -122,12 +122,10 @@ internal static class ClientChannelInstrumentation
         }
     }
 
-    private static IDisposable? SuppressDownstreamInstrumentation()
-    {
-        return WcfInstrumentationActivitySource.Options?.SuppressDownstreamInstrumentation ?? false
+    private static IDisposable? SuppressDownstreamInstrumentation() =>
+        WcfInstrumentationActivitySource.Options?.SuppressDownstreamInstrumentation ?? false
             ? SuppressInstrumentationScope.Begin()
             : null;
-    }
 
     private static ActionMetadata? GetActionMetadata(Message request, string action)
     {