+++ src/OpenTelemetry.Instrumentation.Wcf/Implementation/TelemetryDispatchMessageInspector.cs
@@ -81,7 +81,7 @@ internal class TelemetryDispatchMessageInspector : IDispatchMessageInspector
 
                 if (WcfInstrumentationActivitySource.Options.SetSoapMessageVersion)
                 {
-                    activity.SetTag(WcfInstrumentationConstants.SoapMessageVersionTag, request.Version.ToString());
+                    activity.SetTag(WcfInstrumentationConstants.AttributeSoapMessageVersion, request.Version.ToString());
                 }
 
                 var localAddressUri = channel.LocalAddress?.Uri;
@@ -89,8 +89,8 @@ internal class TelemetryDispatchMessageInspector : IDispatchMessageInspector
                 {
                     activity.SetTag(SemanticConventions.AttributeServerAddress, localAddressUri.Host);
                     activity.SetTag(SemanticConventions.AttributeServerPort, localAddressUri.Port);
-                    activity.SetTag(WcfInstrumentationConstants.WcfChannelSchemeTag, localAddressUri.Scheme);
-                    activity.SetTag(WcfInstrumentationConstants.WcfChannelPathTag, localAddressUri.LocalPath);
+                    activity.SetTag(WcfInstrumentationConstants.AttributeWcfChannelScheme, localAddressUri.Scheme);
+                    activity.SetTag(WcfInstrumentationConstants.AttributeWcfChannelPath, localAddressUri.LocalPath);
                 }
 
                 try
@@ -129,7 +129,7 @@ internal class TelemetryDispatchMessageInspector : IDispatchMessageInspector
                     activity.SetStatus(ActivityStatusCode.Error);
                 }
 
-                activity.SetTag(WcfInstrumentationConstants.SoapReplyActionTag, reply.Headers.Action);
+                activity.SetTag(WcfInstrumentationConstants.AttributeSoapReplyAction, reply.Headers.Action);
                 try
                 {
                     WcfInstrumentationActivitySource.Options!.Enrich?.Invoke(activity, WcfEnrichEventNames.BeforeSendReply, reply);