+++ src/OpenTelemetry.Instrumentation.Wcf/CHANGELOG.md
@@ -13,6 +13,9 @@
 
 * Updated emitted telemetry schema URL to `1.40.0`.
 
+* Updated OpenTelemetry core component version(s) to `1.15.2`.
+  ([#4080](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/4080))
+
 ## 1.15.0-beta.1
 
 Released 2026-Jan-21